feat: concert deletion

+++ src/concerts/concerts.service.ts
@@ -1,4 +1,5 @@
-import { Injectable } from '@nestjs/common';
+import { HttpStatus, Injectable } from '@nestjs/common';
+import { AppError } from '../common/errors/app.error';
 import { CreateConcertDto } from './dto/create-concert.dto';
 import { ConcertRepo } from './concerts.repository';
 
@@ -10,7 +11,20 @@ export class ConcertsService {
     return this.concertRepo.create(dto);
   }
 
-  delete(id: number) {
+  async delete(id: number) {
+    const concert = await this.concertRepo.findById(id);
+
+    if (!concert) {
+      throw new AppError('Concert not found', HttpStatus.NOT_FOUND);
+    }
+
+    if (await this.concertRepo.hasReservationHistory(id)) {
+      throw new AppError(
+        'The concert cannot be removed because it has reservation history',
+        HttpStatus.CONFLICT,
+      );
+    }
+
     return this.concertRepo.delete(id);
   }
 }
